backend/requirements.txt
# ============================================================
#  Backend Dependencies — Cloud Compliance Platform
# ============================================================

# --- Web Framework ---
fastapi==0.111.1
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# --- Auth & Security ---
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
cryptography==42.0.8
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow==10.4.0

# --- Database ---
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2
psycopg2-binary==2.9.9

# --- Caching / Queues ---
redis==5.0.7
celery==5.4.0
kombu==5.3.4

# --- Data Validation ---
pydantic==2.8.2
pydantic-settings==2.3.4
email-validator==2.2.0

# --- Cloud Provider SDKs ---
boto3==1.34.144
botocore==1.34.144
azure-identity==1.17.1
azure-mgmt-resource==23.1.1
azure-mgmt-compute==31.0.0
azure-mgmt-storage==21.1.0
azure-mgmt-network==25.3.0
azure-mgmt-monitor==6.0.2
google-cloud-asset==3.25.0
google-cloud-logging==3.10.0
google-cloud-securitycenter==1.29.0
google-auth==2.32.0
azure-mgmt-sql==3.0.1
azure-mgmt-keyvault==10.3.0
google-cloud-storage==2.17.0
google-cloud-compute==1.19.1
google-cloud-resource-manager==1.12.3
google-api-python-client==2.139.0

# --- OPA Client ---
requests==2.32.3
httpx==0.27.0
aiohttp==3.9.5

# --- Data Pipeline & ML ---
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
pyod==2.0.1
joblib==1.4.2

# --- Reporting ---
jinja2==3.1.4
reportlab==4.2.2

# --- Evidence Storage (MinIO) ---
minio==7.2.7

# --- Alerting ---
slack-sdk==3.31.0
sendgrid==6.11.0

# --- Logging & Monitoring ---
structlog==24.2.0
python-json-logger==2.0.7
prometheus-fastapi-instrumentator==7.0.0

# --- Utilities ---
python-dotenv==1.0.1
pyyaml==6.0.2
click==8.1.7
tenacity==8.5.0
arrow==1.3.0

# --- Dev/Test ---
pytest==8.3.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
factory-boy==3.3.1
faker==26.0.0
ruff==0.5.5
mypy==1.11.1
pip-audit==2.7.3
